Added ActivityItemDetails component and implemented it on all possible items. Restructured folder to have all items on the same level. Cleaned up style files.
