+++ source/ui/ActivityItem/components/items/NFTItem.js
@@ -1,54 +1,44 @@
 import React, { useState } from 'react';
-import NumberFormat from 'react-number-format';
+
 import moment from 'moment';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
-import clsx from 'clsx';
-import {
-  capitalize, IconButton, Typography,
-} from '@material-ui/core';
-import ListIcon from '@material-ui/icons/List';
 
-import ArrowUpRight from '@assets/icons/arrow-up-right.png';
+import { capitalize } from '@material-ui/core';
+
 import UnknownIcon from '@assets/icons/unknown-icon.svg';
+
 import shortAddress from '@shared/utils/short-address';
+
 import { GenericIcon } from '@ui';
 
-import useStyles from '../../styles';
 import ActivityItemDisplay from '../ActivityItemDisplay';
+import ActivityItemDetails from '../ActivityItemDetails';
 
 const getSubtitle = (type, to, from, t) => (({
   SEND: ` · ${t('activity.subtitle.to')}: ${shortAddress(to)}`,
-  BURN: ` · ${t('activity.subtitle.to')}: ${shortAddress(to)}`,
   RECEIVE: ` · ${t('activity.subtitle.from')}: ${shortAddress(from)}`,
 })[type]);
 
 const getAddress = (type, to, from, canisterId) => (
   {
     SEND: to,
-    BURN: to,
     RECEIVE: from,
   }
 )[type] || canisterId || '';
 
-const TokenItem = (props) => {
-  const {
-    type,
-    to,
-    from,
-    amount,
-    value,
-    date,
-    symbol,
-    image,
-    canisterId,
-    details,
-    setOpenDetail,
-    isTransaction,
-    hovering,
-  } = props;
+const NFTItem = ({
+  type,
+  to,
+  from,
+  date,
+  image,
+  canisterId,
+  details,
+  setOpenDetail,
+  hovering,
+}) => {
   const { t } = useTranslation();
-  const classes = useStyles();
 
   const [copied, setCopied] = useState(false);
 
@@ -75,6 +65,7 @@ const TokenItem = (props) => {
       setTooltipText(copyText);
     }, 1500);
   };
+
   return (
     <>
       <ActivityItemDisplay
@@ -83,70 +74,49 @@ const TokenItem = (props) => {
             image={image || UnknownIcon}
             type={type}
           />
-          )}
-        title={`${capitalize(type?.toLowerCase())} ${symbol ?? ''}`}
+        )}
+        title={`${capitalize(type?.toLowerCase())} NFT`}
         subtitle={moment(date).format('MMM Do')}
         tooltip={getSubtitle(type, to, from, t, canisterId)}
         copied={copied}
         tooltipText={tooltipText}
         onCopy={handleClickCopy}
       />
-      <div className={classes.rightContainer}>
-        <div className={classes.amountContainer}>
-          <Typography variant="h5">
-            <NumberFormat value={amount} displayType="text" thousandSeparator="," suffix={` ${symbol}`} decimalScale={5} />
-          </Typography>
-          <Typography variant="subtitle2">
-            <NumberFormat value={value} displayType="text" thousandSeparator="," prefix="$" suffix=" USD" decimalScale={2} />
-          </Typography>
-        </div>
-        <div className={clsx(classes.iconContainer, hovering && classes.iconContainerAnimation)}>
-          {isTransaction ? (
-            <img
-              src={ArrowUpRight}
-            />
-          ) : details && (
-            <IconButton size="small" onClick={() => setOpenDetail(true)} className={classes.detailsIcon}>
-              <ListIcon />
-            </IconButton>
-          )}
-        </div>
-      </div>
+      <ActivityItemDetails
+        main={details?.tokenId?.length > 5 ? shortAddress(details?.tokenId) : `#${details?.tokenId}`}
+        secondary={details?.canisterInfo?.name || shortAddress(canisterId)}
+        hovering={hovering}
+        details={details}
+        setOpenDetail={setOpenDetail}
+      />
     </>
   );
 };
 
-export default TokenItem;
+export default NFTItem;
 
-TokenItem.defaultProps = {
+NFTItem.defaultProps = {
   to: null,
   from: null,
-  amount: null,
-  value: null,
   type: 'PLUG',
-  hovering: false,
+  name: null,
   canisterId: null,
   details: null,
+  hovering: false,
 };
 
-TokenItem.propTypes = {
+NFTItem.propTypes = {
   type: PropTypes.number,
   canisterId: PropTypes.string,
   details: PropTypes.objectOf(PropTypes.any),
   to: PropTypes.string,
   from: PropTypes.string,
-  amount: PropTypes.oneOfType([
-    PropTypes.string,
-    PropTypes.number,
-  ]),
-  value: PropTypes.number,
-  symbol: PropTypes.string.isRequired,
   image: PropTypes.string.isRequired,
   date: PropTypes.oneOfType([
     PropTypes.instanceOf(Date),
     PropTypes.string,
   ]).isRequired,
+  name: PropTypes.string,
   setOpenDetail: PropTypes.func.isRequired,
-  isTransaction: PropTypes.bool.isRequired,
   hovering: PropTypes.bool,
 };
