+++ source/ui/ActivityItem/components/items/SwapItem.js
@@ -3,14 +3,10 @@ import React, { useState } from 'react';
 import moment from 'moment';
 import PropTypes from 'prop-types';
 import { useTranslation } from 'react-i18next';
-import clsx from 'clsx';
-
-import { IconButton, Typography } from '@material-ui/core';
-import ListIcon from '@material-ui/icons/List';
 
 import SwapIcon from '../SwapIcon';
-import useStyles from '../../styles';
 import ActivityItemDisplay from '../ActivityItemDisplay';
+import ActivityItemDetails from '../ActivityItemDetails';
 
 const TokenItem = ({
   date,
@@ -23,8 +19,6 @@ const TokenItem = ({
 
   const { swap } = details?.sonicData || {};
 
-  const classes = useStyles();
-
   return (
     <>
       <ActivityItemDisplay
@@ -34,27 +28,13 @@ const TokenItem = ({
         title={`${t('activity.title.swap')} ${swap?.from?.name} ${t('activity.title.for')} ${swap?.to?.name || t('common.unknownToken')}`}
         subtitle={moment(date).format('MMM Do')}
       />
-      <div className={classes.rightContainer}>
-        <div className={classes.amountContainer}>
-          {/* TODO: Add correct formatting */}
-          <Typography variant="h5">
-            {iconHovered ? 'TODO Hover' : 'TODO'}
-          </Typography>
-        </div>
-        <div className={
-          clsx(
-            classes.iconContainer,
-            hovering && classes.iconContainerAnimation,
-          )
-        }
-        >
-          {details && (
-            <IconButton size="small" onClick={() => setOpenDetail(true)} className={classes.detailsIcon}>
-              <ListIcon />
-            </IconButton>
-          )}
-        </div>
-      </div>
+      <ActivityItemDetails
+        main={iconHovered ? 'TODO Hover' : 'TODO'}
+        secondary=""
+        hovering={hovering}
+        details={details}
+        setOpenDetail={setOpenDetail}
+      />
     </>
   );
 };
